fwk and arch: Fix required MISRA defects rule 10.3
MISRA 10.3 defects relating to files in framework and arch detected
with armv8-m based build, will be addressed in this patch.

+++ arch/arm/armv8-a/include/arch_helpers.h
@@ -509,7 +509,7 @@ DEFINE_RENAME_SYSREG_RW_FUNCS(apiakeylo_el1, APIAKeyLo_EL1)
 
 static inline unsigned int get_current_el(void)
 {
-    return GET_EL(read_CurrentEl());
+    return (unsigned int)GET_EL(read_CurrentEl());
 }
 
 /*
